Added model component to dg. Refactored naming of instrument package to model
