+++ src/main/java/seedu/mtracker/MTracker.java
@@ -4,7 +4,7 @@ import seedu.mtracker.commands.Command;
 import seedu.mtracker.commands.ExitCommand;
 import seedu.mtracker.commands.InvalidCommand;
 import seedu.mtracker.console.InputParser;
-import seedu.mtracker.instrument.InstrumentManager;
+import seedu.mtracker.model.InstrumentManager;
 import seedu.mtracker.ui.TextUi;
 
 import java.util.logging.Level;
