+++ src/main/java/seedu/mtracker/commands/AddCryptoCommand.java
@@ -1,6 +1,6 @@
 package seedu.mtracker.commands;
 
-import seedu.mtracker.instrument.subinstrument.Crypto;
+import seedu.mtracker.model.subinstrument.Crypto;
 import seedu.mtracker.ui.TextUi;
 
 public class AddCryptoCommand extends AddInstrumentCommand {
