+++ src/main/java/seedu/mtracker/commands/AddForexCommand.java
@@ -1,6 +1,6 @@
 package seedu.mtracker.commands;
 
-import seedu.mtracker.instrument.subinstrument.Forex;
+import seedu.mtracker.model.subinstrument.Forex;
 import seedu.mtracker.ui.TextUi;
 
 public class AddForexCommand extends AddInstrumentCommand {
