+++ src/main/java/seedu/mtracker/model/InstrumentManager.java
@@ -1,4 +1,4 @@
-package seedu.mtracker.instrument;
+package seedu.mtracker.model;
 
 import java.util.ArrayList;
